Started implementing the authentication system
Added some tests for the authentication/ authorization system
Added jsonwebtoken library request object in the context to be able to extract the jwt tokens
Removed the ActiveSession object as it is not needed and makes the application more complex for no reason. It also make it possible to avoid calling the database for fetching a session on every call.